requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
paho-mqtt==1.6.1
pymongo==4.5.0
python-dotenv==1.0.0
websockets==12.0
PyJWT==2.8.1
python-multipart==0.0.6